moved picker
